requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
anyascii==0.3.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.11.1
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coverage==7.0.5
crispy-bootstrap5==0.7
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.2
django-allauth==0.52.0
django-bootstrap-v5==1.0.11
django-cache-url==3.4.4
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-dajax==0.9.2
django-dajaxice==0.7
django-elasticsearch-dsl==7.3
django-environ==0.10.0
django-extensions==3.1.5
django-filebrowser==4.0.2
django-filter==22.1
django-grappelli==3.0.4
django-htmx==1.13.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-markdown==0.8.4
django-modelcluster==6.0
django-modeltranslation==0.18.7
django-mptt==0.14.0
django-permissionedforms==0.1
django-rest-framework==0.1.0
django-smart-selects==1.6.0
django-summernote==0.8.20.0
django-taggit==3.1.0
django-templated-mail==1.1.1
django-tinymce==3.5.0
django-treebeard==4.7
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangoajax==3.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
draftjs-exporter==2.1.7
easy-thumbnails==2.8.5
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
environs==9.5.0
et-xmlfile==1.1.0
face-recognition-models==0.3.0
factory-boy==3.2.1
Faker==16.4.0
filetype==1.2.0
frozenlist==1.3.3
graphene==3.2.2
graphene-django==3.0.2
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
import-export==0.3.1
l18n==2021.3
magic-filter==1.0.9
Markdown==3.4.3
MarkupPy==1.14
marshmallow==3.19.0
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.0
Pillow==9.4.0
promise==2.3
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sorl-thumbnail==11.12.1b0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==5.4.0
tablib==3.4.0
telepath==0.3
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.4.0
tzdata==2022.7
Unipath==1.1
urllib3==1.26.14
wagtail==5.0
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
